move "storing a lot of metadata" message
vmn needs to create and store cache on the very first startup. If a
user's library is big, this will take a few minutes. Previously, a
message was shown if, during parsing files, vmn got to the 1000th file.
This can be confusing since it can take some time to get to that file.
Instead, let's print that message while checking the library for new
entries so it displays much sooner and it's clearer to users that vmn is
actually doing something.