rewrite line_split so it abuses the heap less
Strings suck in C (C++ rewrite when). There's a lot of string parsing in
here and one of the cornerstone functions is line_split which is used
everywhere. This is fine, but actually it allocates too much memory and
triggered mysterious bugs with certain mpv config options. I took the
opportunity to rewrite it so it wouldn't do nearly as many malloc/calloc
calls. It also has the nice benefit of making subsequent free calls,
much less painful.

Downside: there's a segfault if you execute an mpv command and then
quit. I'll figure that one out later.