cast correctly in line_split
Unknown if this actually fixes anything. There was a strange, subtle bug
with the malloc call that crashes on occassion. I was able to reproduce
it in gdb by constant constantly queueing a new album. With these casts,
it doesn't seem to crash anymore, but who knows if it actually works.