utils: strip newlines on line_split
This is kind of a major change that possibly may break something, but
hopefuly not. It turns out that line_split would keep a trailing new
line which seems bad and broke another small refactor. cfg_split was a
silly function for reading the config file. line_split accomplishes the
same thing as long as we strip the newline which seems like the right
thing to do.